refactor: Implement onLoadFinished() and onLoaderReset()
onLoadFinished()
The loader framework calls onLoadFinished() when the Contacts Provider returns the results of the query. In this method, put the result Cursor in the SimpleCursorAdapter. This automatically updates the ListView with the search results

onLoaderReset()
invoked when the loader framework detects that the result Cursor contains stale data. Delete the SimpleCursorAdapter reference to the existing Cursor. If you don't, the loader framework will not recycle the Cursor, which causes a memory leak.